Gautam Manu emergent Analytics Intern assignment/requirements.txt
pandas
numpy
google-generativeai
ratelimit
rapidfuzz
python-dotenv